Updated tests to prepare for model changes, tests now exit with status 1 when failing.
Updated tests to reflect the changes to be made to the models.
Several tests should now be failing since the model changes haven't been committed.
Additionally running 'manage.py test' now exits with status 1 when
any of the tests are failing. It previously exited with status 0 regarless
of the tests status which made TravisCI incorrectly report a build as
passing even when several tests failed.